Made the Spear Minigame
Works perfectly on my end, let me know if there's any issues on yours.

If you disable the Image in the canvas, the minigame will not start, becuase I have it set so the game only readies when its alpha is at 0.

The SceneOut() method in BlackFade.cs (attached to the Image in the canvas) is where you can put the scene transition

Changes to the animation for the tentacles can be made through Tentacle.cs

Other scripts added: ThrowSpear.cs, Spear Collisions.cs