Minigame isnt poopy anymore
Animations of arm cocking back now plays when space is being held to indicate throw strength. the stronger the throw the lower the angle. Weak throws will miss overhead

Animations for tentacles are now 30 samples sped up instead of 60. Not sure why, but it decreased overhead and it looks basically the same

Hold space to charge, Victor.